league機能として切り出し
