+++ src/api/services/index.ts
@@ -1,2 +1,1 @@
-export * from './leagueService';
 export * from './gameService';
