+++ src/features/league/api/leagueService.ts
@@ -1,5 +1,5 @@
-import { League, ReqCreateLeague, ResDeleteLeague } from '../../types/league';
-import { apiClient } from '../apiClient';
+import { apiClient } from '../../../api/apiClient';
+import { League, ReqCreateLeague, ResDeleteLeague } from '../../../types/league';
 
 export const fetchLeague = async (id: string, signal?: AbortSignal): Promise<League> => {
   const params = { id: id };
